[posix-app] remove node id from logging
* Simplified posix-app logging
* typedef otLogLevel to signed int so that when dynamic logging and
  reference_device is enabled, there will not be always true error.

+++ include/openthread/platform/logging.h
@@ -110,7 +110,7 @@ extern "C" {
  * This type represents the log level.
  *
  */
-typedef uint8_t otLogLevel;
+typedef int otLogLevel;
 
 /**
  * This enumeration represents log regions.
